pyproject.toml
[tool.poetry]
name = "wordcount"
version = "0.1.0"
description = ""
authors = ["David <davigetto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
PyHamcrest = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
